moving DataClassListJson field to common location for reuse
